requirements.txt
openai
fastapi
black
isort
websockets
pydantic
langchain==0.0.158
uvicorn
jinja2
faiss-cpu
bs4
unstructured
libmagic
rich~=13.3.2
python-dotenv==1.0.0
memphis-py
pymilvus
tiktoken
sqlalchemy==2.0.12